+++ src/common/adapter/fileDialog/FileDialogAdapter.h
@@ -28,7 +28,7 @@ class FileDialogAdapter : public IFileDialogAdapter
 {
 public:
 	explicit FileDialogAdapter() = default;
-	~FileDialogAdapter() = default;
+	~FileDialogAdapter() override = default;
 	QString	getExistingDirectory(QWidget *parent, const QString &title, const QString &directory) override;
 	QString	getOpenFileName(QWidget *parent, const QString &title, const QString &directory) override;
 };